requirements.txt
fastapi==0.121.3
numpy==2.3.5
pandas==2.3.3
pydantic==2.12.4
python-dotenv==1.2.1
Requests==2.32.5
tqdm==4.67.1
uvicorn[standard]==0.24.0
scikit-learn==1.3.2
sentence-transformers==2.2.2